zero-sugar: tune plugin.S per our LPDDR3 initialization script
The most part of imx7d_12x12_arm3 lpddr3 initialization are same to
ours currently done in DCD table.  But a few changes need to be made
to get them match exactly, only except that the following redundant
setting of register 0x30790018 just before CHECK_BITS_SET is dropped.

  DATA 4 0x30790018 0x0000000f

In the meantime, the code for TO 1.1 quirk is dropped, as we are running
on TO 1.3.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>