zero-sugar: drop touch power control initialization
Touch is not needed in U-Boot, so it makes more sense to get kernel
initialize power for touch.  Let's drop the code from U-Boot.  The bonus
point is that we get U-Boot boot up faster now, as the delay for power
control is gone.  If we can drop the digitizer power control later (if
not needed), another 1 second delay can be gone as well, and kernel has
been ready to initialize power for both touch and digitizer anyway.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>