epd_display_init: Specify mmc partition to use
The mmc device and partition settings are global
in uboot.

When other parts of the code uses the mmc, the
splash_init is unable to find the zplash file.

This patch specifies which device and partition
to use, so that it can run independant of what
has been set before.